+++ packages/camera/camera_android_camerax/android/src/main/java/io/flutter/plugins/camerax/CameraControlHostApiImpl.java
@@ -29,7 +29,7 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
   private final InstanceManager instanceManager;
   private final CameraControlProxy proxy;
 
-  /** Proxy for constructors and static method of {@link CameraControl}. */
+  /** Proxy for methods of {@link CameraControl}. */
   @VisibleForTesting
   public static class CameraControlProxy {
     Context context;
@@ -83,6 +83,12 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
             }
 
             public void onFailure(Throwable t) {
+              if (t instanceof CameraControl.OperationCanceledException) {
+                // Operation was canceled due to camera being closed or a new request was submitted, which
+                // is not actionable and should not block a new value from potentially being submitted.
+                result.success(null);
+                return;
+              }
               result.error(t);
             }
           },
@@ -94,11 +100,18 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
      *
      * <p>Will trigger an auto focus action and enable auto focus/auto exposure/auto white balance
      * metering regions.
+     *
+     * <p>Will send a {@link GeneratedCameraXLibrary.Result} with a null result if operation was
+     * canceled.
      */
     public void startFocusAndMetering(
         @NonNull CameraControl cameraControl,
         @NonNull FocusMeteringAction focusMeteringAction,
         @NonNull GeneratedCameraXLibrary.Result<Long> result) {
+      if (context == null) {
+        throw new IllegalStateException("Context must be set to set zoom ratio.");
+      }
+
       ListenableFuture<FocusMeteringResult> focusMeteringResultFuture =
           cameraControl.startFocusAndMetering(focusMeteringAction);
 
@@ -113,6 +126,12 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
             }
 
             public void onFailure(Throwable t) {
+              if (t instanceof CameraControl.OperationCanceledException) {
+                // Operation was canceled due to camera being closed or a new request was submitted, which
+                // is not actionable and should not block a new value from potentially being submitted.
+                result.success(null);
+                return;
+              }
               result.error(t);
             }
           },
@@ -148,6 +167,9 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
      * <p>The exposure compensation value set on the camera must be within the range of {@code
      * ExposureState#getExposureCompensationRange()} for the current {@code ExposureState} for the
      * call to succeed.
+     *
+     * <p>Will send a {@link GeneratedCameraXLibrary.Result} with a null result if operation was
+     * canceled.
      */
     public void setExposureCompensationIndex(
         @NonNull CameraControl cameraControl, @NonNull Long index, @NonNull Result<Long> result) {
@@ -162,6 +184,12 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
             }
 
             public void onFailure(Throwable t) {
+              if (t instanceof CameraControl.OperationCanceledException) {
+                // Operation was canceled due to camera being closed or a new request was submitted, which
+                // is not actionable and should not block a new value from potentially being submitted.
+                result.success(null);
+                return;
+              }
               result.error(t);
             }
           },
@@ -173,6 +201,7 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
    * Constructs an {@link CameraControlHostApiImpl}.
    *
    * @param instanceManager maintains instances stored to communicate with attached Dart objects
+   * @param context {@link Context} used to retrieve {@code Executor}
    */
   public CameraControlHostApiImpl(
       @NonNull BinaryMessenger binaryMessenger,
@@ -185,8 +214,8 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
    * Constructs an {@link CameraControlHostApiImpl}.
    *
    * @param instanceManager maintains instances stored to communicate with attached Dart objects
-   * @param proxy proxy for constructors and static method of {@link CameraControl}
-   * @param context {@link Context} used to retrieve {@code Executor} used to enable torch mode
+   * @param proxy proxy for methods of {@link CameraControl}
+   * @param context {@link Context} used to retrieve {@code Executor}
    */
   @VisibleForTesting
   CameraControlHostApiImpl(
@@ -203,11 +232,11 @@ public class CameraControlHostApiImpl implements CameraControlHostApi {
   }
 
   /**
-   * Sets the context that the {@code ProcessCameraProvider} will use to enable/disable torch mode
-   * and set the zoom ratio.
+   * Sets the context that the {@code CameraControl} will use to enable/disable torch mode and set
+   * the zoom ratio.
    *
-   * <p>If using the camera plugin in an add-to-app context, ensure that a new instance of the
-   * {@code CameraControl} is fetched via {@code #enableTorch} anytime the context changes.
+   * <p>If using the camera plugin in an add-to-app context, ensure that this is called anytime that
+   * the context changes.
    */
   public void setContext(@NonNull Context context) {
     this.proxy.context = context;