+++ packages/in_app_purchase/in_app_purchase_storekit/test/store_kit_wrappers/sk_methodchannel_apis_test.dart
@@ -4,7 +4,6 @@
 
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:in_app_purchase_storekit/src/channel.dart';
 import 'package:in_app_purchase_storekit/src/messages.g.dart';
 import 'package:in_app_purchase_storekit/store_kit_wrappers.dart';
 import '../test_api.g.dart';
@@ -17,9 +16,6 @@ void main() {
 
   setUpAll(() {
     TestInAppPurchaseApi.setup(fakeStoreKitPlatform);
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(
-            SystemChannels.platform, fakeStoreKitPlatform.onMethodCall);
   });
 
   setUp(() {});
@@ -76,10 +72,10 @@ void main() {
     });
 
     test('refreshed receipt', () async {
-      final int receiptCountBefore = fakeStoreKitPlatform.refreshReceipt;
+      final int receiptCountBefore = fakeStoreKitPlatform.refreshReceiptCount;
       await SKRequestMaker().startRefreshReceiptRequest(
           receiptProperties: <String, dynamic>{'isExpired': true});
-      expect(fakeStoreKitPlatform.refreshReceipt, receiptCountBefore + 1);
+      expect(fakeStoreKitPlatform.refreshReceiptCount, receiptCountBefore + 1);
       expect(fakeStoreKitPlatform.refreshReceiptParam,
           <String, dynamic>{'isExpired': true});
     });
@@ -175,9 +171,9 @@ void main() {
     });
 
     test('showPriceConsentIfNeeded should call methodChannel', () async {
-      expect(fakeStoreKitPlatform.showPriceConsentIfNeeded, false);
+      expect(fakeStoreKitPlatform.showPriceConsent, false);
       await SKPaymentQueueWrapper().showPriceConsentIfNeeded();
-      expect(fakeStoreKitPlatform.showPriceConsentIfNeeded, true);
+      expect(fakeStoreKitPlatform.showPriceConsent, true);
     });
   });
 
@@ -192,10 +188,6 @@ void main() {
 }
 
 class FakeStoreKitPlatform implements TestInAppPurchaseApi {
-  FakeStoreKitPlatform() {
-    TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
-        .setMockMethodCallHandler(channel, onMethodCall);
-  }
   // get product request
   List<dynamic> startProductRequestParam = <dynamic>[];
   bool getProductRequestFailTest = false;
@@ -205,7 +197,7 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
   bool getReceiptFailTest = false;
 
   // refresh receipt request
-  int refreshReceipt = 0;
+  int refreshReceiptCount = 0;
   late Map<String, dynamic> refreshReceiptParam;
 
   // payment queue
@@ -217,63 +209,13 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
   bool presentCodeRedemption = false;
 
   // show price consent sheet
-  bool showPriceConsentIfNeeded = false;
+  bool showPriceConsent = false;
 
   // indicate if the payment queue delegate is registered
   bool isPaymentQueueDelegateRegistered = false;
 
   // Listen to purchase updates
   bool? queueIsActive;
-
-  Future<dynamic> onMethodCall(MethodCall call) {
-    switch (call.method) {
-      // request makers
-      case '-[InAppPurchasePlugin startProductRequest:result:]':
-        startProductRequestParam = call.arguments as List<dynamic>;
-        if (getProductRequestFailTest) {
-          return Future<dynamic>.value();
-        }
-        return Future<Map<String, dynamic>>.value(
-            buildProductResponseMap(dummyProductResponseWrapper));
-      case '-[InAppPurchasePlugin refreshReceipt:result:]':
-        refreshReceipt++;
-        refreshReceiptParam = Map.castFrom<dynamic, dynamic, String, dynamic>(
-            call.arguments as Map<dynamic, dynamic>);
-        return Future<void>.sync(() {});
-      // receipt manager
-      case '-[InAppPurchasePlugin retrieveReceiptData:result:]':
-        if (getReceiptFailTest) {
-          throw Exception('some arbitrary error');
-        }
-        return Future<String>.value('receipt data');
-      case '-[InAppPurchasePlugin finishTransaction:result:]':
-        transactionsFinished.add(
-            Map<String, String>.from(call.arguments as Map<dynamic, dynamic>));
-        return Future<void>.sync(() {});
-      case '-[InAppPurchasePlugin restoreTransactions:result:]':
-        applicationNameHasTransactionRestored = call.arguments as String;
-        return Future<void>.sync(() {});
-      case '-[InAppPurchasePlugin presentCodeRedemptionSheet:result:]':
-        presentCodeRedemption = true;
-        return Future<void>.sync(() {});
-      case '-[SKPaymentQueue startObservingTransactionQueue]':
-        queueIsActive = true;
-        return Future<void>.sync(() {});
-      case '-[SKPaymentQueue stopObservingTransactionQueue]':
-        queueIsActive = false;
-        return Future<void>.sync(() {});
-      case '-[SKPaymentQueue registerDelegate]':
-        isPaymentQueueDelegateRegistered = true;
-        return Future<void>.sync(() {});
-      case '-[SKPaymentQueue removeDelegate]':
-        isPaymentQueueDelegateRegistered = false;
-        return Future<void>.sync(() {});
-      case '-[SKPaymentQueue showPriceConsentIfNeeded]':
-        showPriceConsentIfNeeded = true;
-        return Future<void>.sync(() {});
-    }
-    return Future<dynamic>.error('method not mocked');
-  }
 
   @override
   void addPayment(Map<String?, Object?> paymentMap) {
@@ -295,6 +237,73 @@ class FakeStoreKitPlatform implements TestInAppPurchaseApi {
   @override
   List<SKPaymentTransactionMessage?> transactions() =>
       <SKPaymentTransactionMessage>[dummyTransactionMessage];
+
+  @override
+  void finishTransaction(Map<String?, String?> finishMap) {
+    transactionsFinished.add(Map<String, String>.from(finishMap));
+  }
+
+  @override
+  void presentCodeRedemptionSheet() {
+    presentCodeRedemption = true;
+  }
+
+  @override
+  void restoreTransactions(String? applicationUserName) {
+    applicationNameHasTransactionRestored = applicationUserName!;
+  }
+
+  @override
+  Future<SKProductsResponseMessage> startProductRequest(
+      List<String?> productIdentifiers) {
+    startProductRequestParam = productIdentifiers;
+    if (getProductRequestFailTest) {
+      return Future<SKProductsResponseMessage>.value(
+          SKProductsResponseMessage());
+    }
+    return Future<SKProductsResponseMessage>.value(dummyProductResponseMessage);
+  }
+
+  @override
+  void registerPaymentQueueDelegate() {
+    isPaymentQueueDelegateRegistered = true;
+  }
+
+  @override
+  void removePaymentQueueDelegate() {
+    isPaymentQueueDelegateRegistered = false;
+  }
+
+  @override
+  void startObservingPaymentQueue() {
+    queueIsActive = true;
+  }
+
+  @override
+  void stopObservingPaymentQueue() {
+    queueIsActive = false;
+  }
+
+  @override
+  String retrieveReceiptData() {
+    if (getReceiptFailTest) {
+      throw Exception('some arbitrary error');
+    }
+    return 'receipt data';
+  }
+
+  @override
+  Future<void> refreshReceipt({Map<String?, dynamic>? receiptProperties}) {
+    refreshReceiptCount++;
+    refreshReceiptParam =
+        Map.castFrom<dynamic, dynamic, String, dynamic>(receiptProperties!);
+    return Future<void>.sync(() {});
+  }
+
+  @override
+  void showPriceConsentIfNeeded() {
+    showPriceConsent = true;
+  }
 }
 
 class TestPaymentQueueDelegate extends SKPaymentQueueDelegateWrapper {}