+++ packages/url_launcher/url_launcher_platform_interface/lib/link.dart
@@ -86,10 +86,7 @@ abstract class LinkInfo {
 // TODO(ianh): Remove the first argument.
 Future<ByteData> pushRouteNameToFramework(Object? _, String routeName) {
   final Completer<ByteData> completer = Completer<ByteData>();
-  // TODO(chunhtai): remove this ignore and migrate the code
-  // https://github.com/flutter/flutter/issues/124045.
-  // ignore: deprecated_member_use
-  SystemNavigator.routeInformationUpdated(location: routeName);
+  SystemNavigator.routeInformationUpdated(uri: Uri.parse(routeName));
   ui.channelBuffers.push(
     'flutter/navigation',
     _codec.encodeMethodCall(