+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,34 @@
+## 0.5.0+36
+
+* Implements `setExposureMode`.
+
+## 0.5.0+35
+
+* Modifies `CameraInitializedEvent` that is sent when the camera is initialized to indicate that the initial focus
+  and exposure modes are auto and that developers may set focus and exposure points.
+
+## 0.5.0+34
+
+* Implements `setFocusPoint`, `setExposurePoint`, and `setExposureOffset`.
+
+## 0.5.0+33
+
+* Fixes typo in `README.md`.
+* Updates minimum supported SDK version to Flutter 3.13/Dart 3.1.
+
+## 0.5.0+32
+
+* Removes all remaining `unawaited` calls to fix potential race conditions and updates the
+  camera state when video capture starts.
+
+## 0.5.0+31
+
+* Wraps CameraX classes needed to set capture request options, which is needed to implement setting the exposure mode.
+
+## 0.5.0+30
+
+* Adds documentation to clarify how the plugin uses resolution presets as target resolutions for CameraX.
+
 ## 0.5.0+29
 
 * Modifies `buildPreview` to return `Texture` that maps to camera preview, building in the assumption